+++ cmd/destroy.go
@@ -68,6 +68,11 @@ func newDestroyCmd() *cobra.Command {
 				return result.FromError(err)
 			}
 
+			var displayType = display.DisplayProgress
+			if diffDisplay {
+				displayType = display.DisplayDiff
+			}
+
 			opts.Display = display.Options{
 				Color:                cmdutil.GetGlobalColorization(),
 				ShowConfig:           showConfig,
@@ -75,7 +80,7 @@ func newDestroyCmd() *cobra.Command {
 				ShowSameResources:    showSames,
 				SuppressOutputs:      suppressOutputs,
 				IsInteractive:        interactive,
-				DiffDisplay:          diffDisplay,
+				Type:                 displayType,
 				Debug:                debug,
 			}