+++ cmd/up.go
@@ -289,6 +289,11 @@ func newUpCmd() *cobra.Command {
 				return result.FromError(err)
 			}
 
+			var displayType = display.DisplayProgress
+			if diffDisplay {
+				displayType = display.DisplayDiff
+			}
+
 			opts.Display = display.Options{
 				Color:                cmdutil.GetGlobalColorization(),
 				ShowConfig:           showConfig,
@@ -296,7 +301,7 @@ func newUpCmd() *cobra.Command {
 				ShowSameResources:    showSames,
 				SuppressOutputs:      suppressOutputs,
 				IsInteractive:        interactive,
-				DiffDisplay:          diffDisplay,
+				Type:                 displayType,
 				Debug:                debug,
 			}