+++ pkg/backend/display/options.go
@@ -16,6 +16,18 @@ package display
 
 import "github.com/pulumi/pulumi/pkg/diag/colors"
 
+// Type of output to display.
+type Type int
+
+const (
+	// DisplayProgress displays an update as it progresses.
+	DisplayProgress Type = iota
+	// DisplayDiff displays a rich diff.
+	DisplayDiff
+	// DisplayQuery displays query output.
+	DisplayQuery
+)
+
 // Options controls how the output of events are rendered
 type Options struct {
 	Color                colors.Colorization // colorization to apply to events.
@@ -23,8 +35,9 @@ type Options struct {
 	ShowReplacementSteps bool                // true to show the replacement steps in the plan.
 	ShowSameResources    bool                // true to show the resources that aren't updated in addition to updates.
 	SuppressOutputs      bool                // true to suppress output summarization, e.g. if contains sensitive info.
-	SummaryDiff          bool                // If the diff display should be summarized
-	IsInteractive        bool                // If we should display things interactively
-	DiffDisplay          bool                // true if we should display things as a rich diff
+	SummaryDiff          bool                // true if diff display should be summarized.
+	IsInteractive        bool                // true if we should display things interactively.
+	Type                 Type                // type of display (rich diff, progress, or query).
+	JSONDisplay          bool                // true if we should emit the entire diff as JSON.
 	Debug                bool                // true to enable debug output.
 }