+++ pkg/util/ciutil/vars.go
@@ -15,6 +15,7 @@
 package ciutil
 
 import (
+	"fmt"
 	"os"
 )
 
@@ -34,36 +35,75 @@ type Vars struct {
 	BranchName string
 	// CommitMessage is the full message of the Git commit being built.
 	CommitMessage string
+	// PRNumber is the pull-request ID/number in the source control system.
+	PRNumber string
 }
 
 // DetectVars detects and returns the CI variables for the current environment.
+// Not all fields of the `Vars` struct are applicable to every CI system,
+// and may be left blank.
 func DetectVars() Vars {
+	if os.Getenv("PULUMI_DISABLE_CI_DETECTION") != "" {
+		return Vars{Name: ""}
+	}
+
 	v := Vars{Name: DetectSystem()}
-	// All CI systems have a name (that's how we know we're in a CI system). After detecting one, we will
-	// try to detect some additional CI-specific metadata that the CLI will use. It's okay if we can't,
-	// we'll just have reduced functionality for our various CI integrations.
+
 	switch v.Name {
-	case GitLab:
-		// See https://docs.gitlab.com/ee/ci/variables/.
-		v.BuildID = os.Getenv("CI_JOB_ID")
-		v.BuildURL = os.Getenv("CI_JOB_URL")
-		v.SHA = os.Getenv("CI_COMMIT_SHA")
-		v.BranchName = os.Getenv("CI_COMMIT_REF_NAME")
-		v.CommitMessage = os.Getenv("CI_COMMIT_MESSAGE")
-	case Travis:
-		// See https://docs.travis-ci.com/user/environment-variables/.
-		// NOTE: Travis doesn't set a build URL in its environment, https://github.com/travis-ci/travis-ci/issues/8935.
-		v.BuildID = os.Getenv("TRAVIS_JOB_ID")
-		v.BuildType = os.Getenv("TRAVIS_EVENT_TYPE")
-		v.SHA = os.Getenv("TRAVIS_PULL_REQUEST_SHA")
-		v.BranchName = os.Getenv("TRAVIS_BRANCH")
-		v.CommitMessage = os.Getenv("TRAVIS_COMMIT_MESSAGE")
+	case AzurePipelines:
+		// See:
+		// https://docs.microsoft.com/en-us/azure/devops/pipelines/build/variables?view=azure-devops&tabs=yaml#build-variables
+		v.BuildID = os.Getenv("BUILD_BUILDID")
+		v.BuildType = os.Getenv("BUILD_REASON")
+		v.SHA = os.Getenv("BUILD_SOURCEVERSION")
+		v.BranchName = os.Getenv("BUILD_SOURCEBRANCHNAME")
+		v.CommitMessage = os.Getenv("BUILD_SOURCEVERSIONMESSAGE")
+		// Azure Pipelines can be connected to external repos.
+		// So we check if the provider is GitHub, then we use
+		// `SYSTEM_PULLREQUEST_PULLREQUESTNUMBER` instead of `SYSTEM_PULLREQUEST_PULLREQUESTID`.
+		// The PR ID/number only applies to Git repos.
+		vcsProvider := os.Getenv("BUILD_REPOSITORY_PROVIDER")
+		switch vcsProvider {
+		case "TfsGit":
+			orgURI := os.Getenv("SYSTEM_PULLREQUEST_TEAMFOUNDATIONCOLLECTIONURI")
+			projectName := os.Getenv("SYSTEM_TEAMPROJECT")
+			v.BuildURL = fmt.Sprintf("%v/%v/_build/results?buildId=%v", orgURI, projectName, v.BuildID)
+			// TfsGit is a git repo hosted on Azure DevOps.
+			v.PRNumber = os.Getenv("SYSTEM_PULLREQUEST_PULLREQUESTID")
+		case "GitHub":
+			// GitHub is a git repo hosted on GitHub.
+			v.PRNumber = os.Getenv("SYSTEM_PULLREQUEST_PULLREQUESTNUMBER")
+		}
 	case CircleCI:
 		// See: https://circleci.com/docs/2.0/env-vars/
 		v.BuildID = os.Getenv("CIRCLE_BUILD_NUM")
 		v.BuildURL = os.Getenv("CIRCLE_BUILD_URL")
 		v.SHA = os.Getenv("CIRCLE_SHA1")
 		v.BranchName = os.Getenv("CIRCLE_BRANCH")
+	case GenericCI:
+		v.Name = System(os.Getenv("PULUMI_CI_SYSTEM"))
+		v.BuildID = os.Getenv("PULUMI_CI_BUILD_ID")
+		v.BuildType = os.Getenv("PULUMI_CI_BUILD_TYPE")
+		v.BuildURL = os.Getenv("PULUMI_CI_BUILD_URL")
+		v.SHA = os.Getenv("PULUMI_CI_PULL_REQUEST_SHA")
+	case GitLab:
+		// See https://docs.gitlab.com/ee/ci/variables/.
+		v.BuildID = os.Getenv("CI_JOB_ID")
+		v.BuildType = os.Getenv("CI_PIPELINE_SOURCE")
+		v.BuildURL = os.Getenv("CI_JOB_URL")
+		v.SHA = os.Getenv("CI_COMMIT_SHA")
+		v.BranchName = os.Getenv("CI_COMMIT_REF_NAME")
+		v.CommitMessage = os.Getenv("CI_COMMIT_MESSAGE")
+		v.PRNumber = os.Getenv("CI_MERGE_REQUEST_ID")
+	case Travis:
+		// See https://docs.travis-ci.com/user/environment-variables/.
+		v.BuildID = os.Getenv("TRAVIS_JOB_ID")
+		v.BuildType = os.Getenv("TRAVIS_EVENT_TYPE")
+		v.BuildURL = os.Getenv("TRAVIS_BUILD_WEB_URL")
+		v.SHA = os.Getenv("TRAVIS_PULL_REQUEST_SHA")
+		v.BranchName = os.Getenv("TRAVIS_BRANCH")
+		v.CommitMessage = os.Getenv("TRAVIS_COMMIT_MESSAGE")
+		v.PRNumber = os.Getenv("TRAVIS_PULL_REQUEST")
 	}
 	return v
 }